refactor(wallet-mobile): collocated api errors
